Fixed input event buffer issues
Queue doesn't properly act as a ring buffer had to roll my own.
I want to make a allocated ring buffer as well...

+++ code/grime/stack_fixed.odin
@@ -12,7 +12,7 @@ stack_clear :: #force_inline proc ( using stack : ^StackFixed( $Type, $Size)) {
 }
 
 stack_push :: #force_inline proc( using stack : ^ StackFixed( $ Type, $ Size ), value : Type ) {
-	assert( idx < len( items ), "Attempted to push on a full stack" )
+	assert( idx < u32(len( items )), "Attempted to push on a full stack" )
 
 	items[ idx ] = value
 	idx += 1
